03_system/01_url_shortener/python_app/requirements.txt
FASTAPI
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
asyncpg
greenlet
pytest
pytest-asyncio
httpx
aiosqlite
redis
gunicorn

